+++ Source/BuildPackages/Templates/Utils_contains.pas
@@ -1,0 +1,12 @@
+contains
+  NovusUtilities in '..\..\core\Utilities\NovusUtilities.pas',
+  NovusWindows in '..\..\Core\Utilities\NovusWindows.pas',
+  NovusVersionUtils in '..\..\Core\Utilities\NovusVersionUtils.pas',
+  NovusDateStringUtils in '..\..\Core\Utilities\NovusDateStringUtils.pas',
+  NovusDateUtils in '..\..\Core\Utilities\NovusDateUtils.pas',
+  NovusStringUtils in '..\..\Core\Utilities\NovusStringUtils.pas',
+  NovusFileUtils in '..\..\Core\Utilities\NovusFileUtils.pas',
+  NovusNumUtils in '..\..\Core\Utilities\NovusNumUtils.pas',
+  NovusDateDiffUtil in '..\..\Core\Utilities\NovusDateDiffUtil.pas',
+  NovusGUIDEx in '..\..\Core\Utilities\NovusGUIDEx.pas',
+  NovusVariants in '..\..\Core\Utilities\NovusVariants.pas';