+++ Source/BuildPackages/Templates/Utils_requires.pas
@@ -1,0 +1,7 @@
+requires
+  rtl,
+  vcl,
+  xmlrtl,
+  dbrtl,
+  soaprtl,
+  inet;